refactor(auth): src/app/features를 src/features로 위치 변경
