+++ src/app/auth/signup/email/page.tsx
@@ -1,6 +1,6 @@
 ' use client';
 
-import EmailSignup from '@/app/features/auth/signup/email/components/email-signup';
+import EmailSignup from '@/features/auth/signup/email/components/email-signup';
 
 const EmailSignupPage = () => {
   return (
